FIX: Resolve compilation errors for Pilot Audit workflow
- Fix uint256 to uint64 conversion in PayoutDistributor.sol:94
- Add missing events to IComplianceRegistry interface
- Fix event emission references in ComplianceRegistry.t.sol tests
- Ensure proper event visibility for testing

Ready for workflow to run successfully now.

+++ src/PayoutDistributor.sol
@@ -91,7 +91,7 @@ contract PayoutDistributor is IPayoutDistributor, Ownable, ReentrancyGuard {
         // Store snapshot data
         snapshots[snapshotId] = Snapshot({
             totalSupply: totalSupply,
-            blockNumber: block.number,
+            blockNumber: uint64(block.number),
             timestamp: uint64(block.timestamp)
         });
